+++ Questionable.Model/Questing/EInteractionType.cs
@@ -45,4 +45,6 @@ public enum EInteractionType
     AcceptQuest,
     CompleteQuest,
     Fish,
+    Macro,
+    CleanUpOtherQuests,
 }