+++ Questionable.Model/Questing/SkipStepConditions.cs
@@ -30,6 +30,8 @@ public sealed class SkipStepConditions
     public NearPositionCondition? NearPosition { get; set; }
     public NearPositionCondition? NotNearPosition { get; set; }
     public EExtraSkipCondition? ExtraCondition { get; set; }
+    public List<uint> DutiesUnlocked { get; set; } = [];
+    public List<uint> DutiesCompleted { get; set; } = [];
 
     public bool HasSkipConditions()
     {
@@ -50,9 +52,34 @@ public sealed class SkipStepConditions
                AetheryteUnlocked != null ||
                NearPosition != null ||
                NotNearPosition != null ||
-               ExtraCondition != null;
+               ExtraCondition != null ||
+               DutiesUnlocked.Count > 0 ||
+               DutiesCompleted.Count > 0;
     }
 
-    public override string ToString() =>
-        $"{nameof(Never)}: {Never}, {nameof(CompletionQuestVariablesFlags)}: {CompletionQuestVariablesFlags}, {nameof(Flying)}: {Flying}, {nameof(Chocobo)}: {Chocobo}, {nameof(Diving)}: {Diving}, {nameof(NotTargetable)}: {NotTargetable}, {nameof(InTerritory)}: {string.Join(" ", InTerritory)}, {nameof(NotInTerritory)}: {string.Join(" ", NotInTerritory)}, {nameof(Item)}: {Item}, {nameof(QuestsAccepted)}: {string.Join(" ", QuestsAccepted)}, {nameof(QuestsCompleted)}: {string.Join(" ", QuestsCompleted)}, {nameof(NotNamePlateIconId)}: {string.Join(" ", NotNamePlateIconId)}, {nameof(NearPosition)}: {NearPosition}, {nameof(ExtraCondition)}: {ExtraCondition}";
+    public override string ToString()
+    {
+        var tmp = new string?[] {
+            Never ? $"{nameof(Never)}: {Never}" : null,
+            CompletionQuestVariablesFlags.Any(x => x != null) ? $"{nameof(CompletionQuestVariablesFlags)}: {string.Join(" ", CompletionQuestVariablesFlags)}" : null,
+            Flying != null ? $"{nameof(Flying)}: {Flying}" : null,
+            Chocobo != null ? $"{nameof(Chocobo)}: {Chocobo}" : null,
+            Diving != null ? $"{nameof(Diving)}: {Diving}" : null,
+            NotTargetable ? $"{nameof(NotTargetable)}: {NotTargetable}" : null,
+            InTerritory.Count > 0 ? $"{nameof(InTerritory)}: {string.Join(" ", InTerritory)}" : null,
+            NotInTerritory.Count > 0 ? $"{nameof(NotInTerritory)}: {string.Join(" ", NotInTerritory)}" : null,
+            Item != null ? $"{nameof(Item)}: {Item}" : null,
+            QuestsAccepted.Count > 0 ? $"{nameof(QuestsAccepted)}: {string.Join(" ", QuestsAccepted)}" : null,
+            QuestsCompleted.Count > 0 ? $"{nameof(QuestsCompleted)}: {string.Join(" ", QuestsCompleted)}" : null,
+            NotNamePlateIconId.Count > 0 ? $"{nameof(NotNamePlateIconId)}: {string.Join(" ", NotNamePlateIconId)}" : null,
+            AetheryteLocked != null ? $"{nameof(AetheryteLocked)}: {AetheryteLocked}" : null,
+            AetheryteUnlocked != null ? $"{nameof(AetheryteUnlocked)}: {AetheryteUnlocked}" : null,
+            NearPosition != null ? $"{nameof(NearPosition)}: {NearPosition}" : null,
+            NotNearPosition != null ? $"{nameof(NotNearPosition)}: {NotNearPosition}" : null,
+            ExtraCondition != null ? $"{nameof(ExtraCondition)}: {ExtraCondition}" : null,
+            DutiesUnlocked.Count > 0 ? $"{nameof(DutiesUnlocked)}: {string.Join(" ", DutiesUnlocked)}" : null,
+            DutiesCompleted.Count > 0 ? $"{nameof(DutiesCompleted)}: {string.Join(" ", DutiesCompleted)}" : null,
+        };
+        return string.Join(", ", tmp.Where(x => x != null));
+    }
 }