+++ Questionable/Windows/QuestComponents/ActiveQuestComponent.cs
@@ -358,7 +358,7 @@ internal sealed partial class ActiveQuestComponent
                 }
 
                 bool showStopClock = hasLevelCondition || hasCompleteQuestConditions || hasAcceptQuestConditions || preventQuestCompletion;
-                bool showPriorityCrystal = anyAvailable || anyUnavailable;
+                bool showPriorityCrystal = true; //anyAvailable || anyUnavailable;
                 if (showStopClock || showPriorityCrystal)
                     ImGui.SameLine();
 
@@ -458,7 +458,7 @@ internal sealed partial class ActiveQuestComponent
                 {
                     if (showStopClock)
                         ImGui.SameLine();
-                    ImGui.TextColored(QstTheme.Amber, SeIconChar.Hyadelyn.ToIconString());
+                    ImGui.TextColored(anyAvailable || anyUnavailable ? QstTheme.Amber : QstTheme.Text, SeIconChar.Hyadelyn.ToIconString());
                     if (ImGui.IsItemHovered())
                     {
                         List<ElementId> availablePriorityQuests = priorityQuests